refactor(mutcat): split _mutcat.py into codebook/classify/count package
