+++ python/genoray/_mutcat/__init__.py
@@ -1,0 +1,79 @@
+"""Mutation-catalogue codebooks, vectorized classifiers, and count matrices.
+
+Package-internal API (no ``import genoray`` reach). Re-exported flat so existing
+``from genoray._mutcat import X`` sites are unaffected by the SP-7 split.
+"""
+
+from .classify import (
+    _BASE2IDX,
+    _COMP,
+    _ID1_CODE,
+    _IDM_CODE,
+    _IDR_CODE,
+    _REF_MISMATCH,
+    _SUB_LUT,
+    _build_dbs_table,
+    _comp,
+    _dbs78_codes,
+    _id83_codes_for_contig,
+    _sbs96_codes,
+    build_entry_codes,
+    classify_dbs78,
+    classify_variants,
+)
+from .codebook import (
+    _BASES,
+    _SBS_SUBS,
+    DBS78,
+    DBS78_INDEX,
+    DBS78_OFFSET,
+    ID83,
+    ID83_INDEX,
+    ID83_OFFSET,
+    Kind,
+    MUTCAT_VERSION,
+    N_CODES,
+    SBS96,
+    SBS96_INDEX,
+    SBS96_OFFSET,
+    Sentinel,
+    code_ranges,
+    labels,
+)
+from .count import count_matrix
+
+__all__ = [  # match the Step 1 inventory exactly
+    "DBS78",
+    "DBS78_INDEX",
+    "DBS78_OFFSET",
+    "ID83",
+    "ID83_INDEX",
+    "ID83_OFFSET",
+    "Kind",
+    "MUTCAT_VERSION",
+    "N_CODES",
+    "SBS96",
+    "SBS96_INDEX",
+    "SBS96_OFFSET",
+    "Sentinel",
+    "_BASE2IDX",
+    "_BASES",
+    "_COMP",
+    "_ID1_CODE",
+    "_IDM_CODE",
+    "_IDR_CODE",
+    "_REF_MISMATCH",
+    "_SBS_SUBS",
+    "_SUB_LUT",
+    "_build_dbs_table",
+    "_comp",
+    "_dbs78_codes",
+    "_id83_codes_for_contig",
+    "_sbs96_codes",
+    "build_entry_codes",
+    "classify_dbs78",
+    "classify_variants",
+    "code_ranges",
+    "count_matrix",
+    "labels",
+]
